replaced authentication regexes

+++ client/src/renderer/services/AdminService.js
@@ -25,7 +25,7 @@ export default {
   deleteAdmin(adminId, imgPath) {
     return Api().delete(`admin/${adminId}`, {data: { imgPath: imgPath }})
   },
-  deleteUser(userId, imgPath) {
-    return Api().delete(`admin/user/${userId}`, {data: { imgPath: imgPath }})
+  deleteUser(adminId, userId, imgPath) {
+    return Api().delete(`admin/${adminId}/user/${userId}`, {data: { imgPath: imgPath }})
   }
 }
